#17 feat(eval): refactor
